Change how layer and factions are discovered
Since the data returned by ShowServerInfo is delayed up to 30 seconds, it
can not be relied on to determine current layer and faction. Especially not
at the start of rounds. Instead, use ShowCurrentMap and ShowNextMap for
this. They return the actual current state of the server.

Also, since the same layer object can be returned for both the current map
and the next, we cannot store team information there. Thus, a new team
object is returned in addition to the layer object. This requires parsing
and returning additional fields from rcon.getCurrentMap() and
rcon.getNextMap().